+++ examples/ahpy_setuptools/ahpy_external.h
@@ -10,6 +10,7 @@ signed char ahpy_external_byte(signed char value);
 int ahpy_external_byte_calls(void);
 double ahpy_external_scale(double value, double factor);
 long long ahpy_external_nogil_probe(void);
+long long ahpy_external_nogil_advance(long long amount);
 long long ahpy_external_nogil_probe_calls(void);
 
 #endif